+++ bindings/nodejs/test/client/offlineSigningExamples.spec.ts
@@ -67,15 +67,16 @@ describe('Offline signing examples', () => {
     // transaction tests disabled for workflows, because they fail if we don't have funds
     it.skip('sends a transaction', async () => {
         // Send block with the signed transaction as a payload
-        const block = await onlineClient.postBlockPayload(
+        const blockIdAndBlock = await onlineClient.postBlockPayload(
             // Imported JSON is typed with literal types
             signedTransactionJson as unknown as PayloadTypes,
         );
 
-        expect(block.payload).toBeDefined();
+        expect(blockIdAndBlock[1].payload).toBeDefined();
 
-        const blockId = await onlineClient.blockId(block);
+        const blockId = await onlineClient.blockId(blockIdAndBlock[1]);
 
+        expect(blockId).toBe(blockIdAndBlock[0]);
         expect(blockId).toBeValidBlockId;
     });
 });