+++ CHANGELOG.md
@@ -21,9 +21,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased - YYYY-MM-DD
 
+### Added
+
+- `NetworkInfoDto`;
+
 ### Changed
 
 - Update dependencies;
+- Return `NetworkInfoDto` instead of `NetworkInfo` in message_interface;
 
 ## 2.0.1-rc.1 - 2022-09-28